+++ src/inet/linklayer/ext/Ext.cc
@@ -274,7 +274,7 @@ void Ext::handleMessage(cMessage *msg)
             addr.sin_addr.s_addr = htonl(ipv4Header->getDestAddress().getInt());
             auto bytesChunk = packet->peekAllAsBytes();
             size_t packetLength = bytesChunk->copyToBuffer(buffer, sizeof(buffer));
-            ASSERT(packetLength == packet->getByteLength());
+            ASSERT(packetLength == (size_t)packet->getByteLength());
 
             EV << "Delivering an IPv4 packet from "
                << ipv4Header->getSrcAddress()
